correct existing comments
