+++ src/utils/getFilteredTodos.ts
@@ -2,13 +2,15 @@ import { Todo } from '../types/todo/Todo';
 import { StatusFilter } from '../enums/statusFilter';
 
 export function getFilteredTodos(todos: Todo[], statusFilter: StatusFilter) {
+  const filteredTodos = [...todos];
+
   switch (statusFilter) {
     case StatusFilter.Active:
-      return [...todos].filter(todo => !todo.completed);
+      return filteredTodos.filter(todo => !todo.completed);
     case StatusFilter.Completed:
-      return [...todos].filter(todo => todo.completed);
+      return filteredTodos.filter(todo => todo.completed);
     case StatusFilter.All:
     default:
-      return [...todos];
+      return filteredTodos;
   }
 }
